+++ src/api/address-name-service.js
@@ -14,11 +14,8 @@ module.exports = {
   async resolveAddress(address) {
     try {
       if (isUndefined(address)) throw TypeError('address is not defined');
-      const { data } = await axios.get(`${getNakaBaseUrl()}/resolve-address`, {
-        params: {
-          address,
-          apikey: CONFIG.NAKABASE_API_KEY,
-        },
+      const { data } = await axios.post(`${getNakaBaseUrl()}/${CONFIG.NAKABASE_API_KEY}/resolve-address`, {
+        address,
       });
       return data;
     } catch (err) {