+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [#293] - Fix Event sync parser tests
 - [#294] - Update readme for copying data
 - [#296] - Setup Nakabase API key for Docker
+- [#299] - Change address name service api to use post
 - [#300] - Fix http provider error
 - [#301] - Fix Migration 2 error on mainnet